+++ kubejs/client_scripts/tooltips.js
@@ -38,9 +38,9 @@ ItemEvents.tooltip(event => {
     event.addAdvanced('spelunkery:portal_fluid_bottle', (item, advanced, text) => {
         text.remove(1);
     });
-    event.addAdvanced('minecraft:respawn_anchor', (item, advanced, text) => {
-        text.remove(1);
-    });
+    // event.addAdvanced('minecraft:respawn_anchor', (item, advanced, text) => {
+    //     text.remove(1);
+    // });
     event.addAdvanced('minecraft:redstone', (item, advanced, text) => {
         text.remove(1);
     });